Use stdint.h

+++ src/food.h
@@ -2,18 +2,19 @@
 #define FOOD_H
 
 #include <curses.h>
+#include <stdint.h>
 
 /**
  * Food parameters
  */
 typedef struct food_param
 {
-    int x;              /**< x coordinate for food */
-    int y;              /**< y coordinate for food */
-    int food_width;     /**< The width of the game window */
-    int food_height;    /**< The height of the game window */
-    int type;           /**< What type of food. TODO */
-    int isFirst;        /**< True if the food appears for the first time. */
+    int32_t x;              /**< x coordinate for food */
+    int32_t y;              /**< y coordinate for food */
+    int32_t food_width;     /**< The width of the game window */
+    int32_t food_height;    /**< The height of the game window */
+    int32_t type;           /**< What type of food. TODO */
+    int8_t isFirst;         /**< True if the food appears for the first time. */
 } food_param_t;
 
 /**
